+++ src/modules/voxel_rt/TraverseRayPipeline.zig
@@ -149,7 +149,7 @@ pub fn deinit(self: TraverseRayPipeline, ctx: Context) void {
     ctx.destroyPipeline(self.pipeline);
 }
 
-pub fn appendPipelineCommands(self: TraverseRayPipeline, ctx: Context, bounce_index: u32, command_buffer: vk.CommandBuffer) void {
+pub fn appendPipelineCommands(self: TraverseRayPipeline, ctx: Context, bounce_index: usize, command_buffer: vk.CommandBuffer) void {
     const zone = tracy.ZoneN(@src(), @typeName(TraverseRayPipeline) ++ " " ++ @src().fn_name);
     defer zone.End();
